rawger/{Components, NovelDetails}: Add Definer and AppNav to NovelDetails, code cleanup
Hooks up NovelDetails to Definer and AppNavigation. Deletes Paginator and DefinableText code.
