+++ Reed/NovelDetails/viewmodels/NovelDetailsViewModel.swift
@@ -7,7 +7,6 @@
 //
 
 import CoreData
-import SwiftUI
 import SwiftyNarou
 
 class NovelDetailsViewModel: ObservableObject {
@@ -19,10 +18,9 @@ class NovelDetailsViewModel: ObservableObject {
     @Published var isLibraryDataLoading: Bool = true
     @Published var isNovelSynopsisProcessing: Bool = true
     @Published var isFavorite: Bool = false
-    @Published var novelSynopsis = NSMutableAttributedString()
+    @Published var novelSynopsis: ProcessedContent?
     @Published var novelSynopsisHeight = DefinerConstants.CONTENT_HEIGHT
     var historyEntry: HistoryEntry?
-    var tokens = [Token]()
     
     init(persistentContainer: NSPersistentContainer, ncode: String) {
         self.persistentContainer = persistentContainer
@@ -35,20 +33,7 @@ class NovelDetailsViewModel: ObservableObject {
         // Fetch metadata from Narou.
         model.fetchNovelDetails { novelData in
             self.novelData = novelData
-            let tokenizer = Tokenizer()
-            self.tokens = tokenizer.tokenize(novelData?.synopsis ?? "")
-//            self.novelSynopsis = (novelData?.synopsis?.trimmingCharacters(in: ["\n"]) ?? "").annotateWithFurigana(tokens: self.tokens)
-            //calculations
-            let testDefinableTextView = DefinableTextView(
-                frame: CGRect(
-                    x: 0,
-                    y: 0,
-                    width: DefinerConstants.CONTENT_WIDTH,
-                    height: DefinerConstants.NOVEL_SYNOPSIS_PRELOAD_HEIGHT
-                ),
-                content: self.novelSynopsis
-            )
-            self.novelSynopsisHeight = (testDefinableTextView.getLinesYCoordinates().last ?? DefinerConstants.CONTENT_HEIGHT) + testDefinableTextView.font.pointSize
+            self.novelSynopsis = ProcessedContent(content: novelData?.synopsis ?? "")
             self.isNovelSynopsisProcessing = false
         }
         
@@ -107,22 +92,6 @@ class NovelDetailsViewModel: ObservableObject {
             subgenre: novelSubgenre!.rawValue,
             isFavorite: false
         )
-    }
-    
-    func getToken(l: Int, r: Int, x: Int) -> Token? {
-        var i = l
-        var j = r
-        while j >= i {
-            let mid = i + (j - i) / 2
-            if tokens[mid].range.lowerBound <= x && tokens[mid].range.upperBound > x {
-                return tokens[mid]
-            } else if tokens[mid].range.lowerBound > x {
-                j = mid - 1
-            } else {
-                i = mid + 1
-            }
-        }
-        return nil
     }
 }
